move open to its own area
open :wink: it up for use in UserReader
